LPS-176902 (portal-impl) - remove default service account during company deletion

+++ portal-impl/src/com/liferay/portal/service/impl/CompanyLocalServiceImpl.java
@@ -1363,7 +1363,7 @@ public class CompanyLocalServiceImpl extends CompanyLocalServiceBaseImpl {
 			userActionableDynamicQuery.setCompanyId(companyId);
 			userActionableDynamicQuery.setPerformActionMethod(
 				(User user) -> {
-					if (!user.isDefaultUser()) {
+					if (!(user.isDefaultUser() && user.isRegularUser())) {
 						_userLocalService.deleteUser(user.getUserId());
 					}
 				});
